refactor: split graph indexers into indexers.ts — index.ts is now a barrel (#671, increment 3)
Final increment of the graph/index.ts decomposition. Move the indexing/write
layer out into graph/indexers.ts; index.ts becomes init/lifecycle + the
mutation API + the public barrel. Bodies moved VERBATIM; only imports changed.

- graph/indexers.ts (new, 1,231) — indexNote, indexSource/indexSourceBody,
  indexExcerpt, indexCsvTable/unindexCsvTable/unindexAllCsvTables, indexTable,
  indexPythonFile/indexTurtleFile/indexCsvFile, removeNote/removeSource/
  removeExcerpt, excerptIdsForSource, indexAllNotes + walkAndIndex* (internal),
  addOntologyToStore, detectHeadingRename, parse{Source,Excerpt}IdFromPath, and
  the indexer-only helpers (resolveLinkTarget/resolveTargetByAlias/
  rebuildAliasMap, the frontmatter helpers, ensure*, injectPrefixes,
  extractHeadingsFromContent, fileMtimeIso, xsdForDuckDbType). Types
  CsvTableColumn/CsvTableShape/HeadingRenameCandidate. Imports ./state,
  ./write-guard (checkLLMWriteGuard), ./queries (findNotesLinkingToAnchorImpl)
  — never ./index.
- graph/index.ts (1,387 → 201) — disposeProject, config helpers, initGraph
  (imports addOntologyToStore from ./indexers), persist/serialize/export,
  parseIntoStore/removeMatchingTriples, and re-exports the public indexer API +
  types so external `import * as graph` callers are unchanged.

Import graph is acyclic: state (leaf) ← queries ← indexers ← index.

Verified: pnpm lint (0 errors); the ~36 graph tests (34 files / 324 tests) pass
— the cycle/init-order net; full suite (2845 passed); e2e boot smoke. Confirmed
no ./index import in state/queries/indexers.
